windows: Remove XDG dependency.
Windows doesn't benefit from XDG, and it doesn't install nicely like git dependencies.
